Convert png images to url data to avoid taint the canvas
If we load a png image from the local filesystem, the canvas is tainted
and we can't save the content as a image, and we need that.
The solution is convert all the images to urldata format,
then is loaded as javascript code without problems.
A marginal cost is a bigger space neeed to store the images.