Add a css to use in the tablet when in portrait orientation
In the tablet in use in UY, when the tablet is in portrait orientation,
the browser resolution is 980x1594, compared to the 1024x552
in landscape orientation. Then the toolbar and buttons are very tiny.
To solve that, use a new css sugar-300dpi.css, generated using lessc
and a sugar-300dpi.less.
There are a manual change after he css is generated.
The font-size is set to 18pt.
The text in the buttons in the alert are still tiny, need research.