When a image is dragged, disable the play audio button
The name of the audio file is taken from the images in the martrix
(like the names of the pdf and images). If a audio file was recorded,
the play button is enabled, but if a image is dragged,
the file name will not be the same, then we need disable the play button,
and let the user save a new audio file.